2차 수정본
자잘한 오류들 고쳤고, 일단은 여러판 가능한 꼴로 적용 했습니다. 실행되는 2차 결과본 미리 올려둡니다.

+++ ComboManager.cpp
@@ -5,7 +5,7 @@ ComboManager::ComboManager()
 {
 }
 
-int ComboManager::updateByClear(int clearedLines, DWORD now)
+int ComboManager::updateByClear(int clearedLines, ULONGLONG now)
 {
     if (clearedLines <= 0)
     {
